cast ints to float to avoid integer rounding

+++ src/Perpetuum/Zones/Terrains/Materials/Plants/PlantHandler.cs
@@ -22,7 +22,7 @@ namespace Perpetuum.Zones.Terrains.Materials.Plants
 
         private const int AREA_SIZE = 32;
         // Number of hours a full pass of plant regen should take
-        private readonly int FULL_PLANT_REGEN_PASS = 8;
+        private readonly float FULL_PLANT_REGEN_PASS = 8f;
 
         private readonly IZone _zone;
         private readonly TimeSpan _natureSleepAmount = TimeSpan.FromSeconds(7);
@@ -36,7 +36,7 @@ namespace Perpetuum.Zones.Terrains.Materials.Plants
         private int _areaDoneY;
         private readonly int _areaAmount;
         private bool _stopSignal;
-        private int _total_area;
+        private float _total_area;
 
         public Area WorkArea { private get; set; }
 
@@ -44,7 +44,7 @@ namespace Perpetuum.Zones.Terrains.Materials.Plants
         {
             _zone = zone;
             _areaAmount = zone.Size.Width / AREA_SIZE; //the amount of areas
-            _total_area = (zone.Size.Width / AREA_SIZE) * (zone.Size.Height / AREA_SIZE);
+            _total_area = ((float)zone.Size.Width / (float)AREA_SIZE) * ((float)zone.Size.Height / (float)AREA_SIZE);
             _plantsTimer = new IntervalTimer(TimeSpan.FromHours(FULL_PLANT_REGEN_PASS / _total_area));
             WorkArea = zone.Size.ToArea();
